Clan bank is now a 64-bit int.  Modifications to clan are now updated in database.

+++ src/include/clan.h
@@ -59,7 +59,7 @@ struct room_list_elem {
 
 struct clan_data {
 	int number;
-	int bank_account;
+	long long int bank_account;
 	byte top_rank;
 	long owner;
 	int flags;
